requirements.txt
requests>=2.32.5
beautifulsoup4>=4.12.3
lxml>=5.3.0
openpyxl>=3.1.2
openai>=2.0.0
python-dotenv>=1.0.1
fake-useragent>=1.4.0
